lesson10/requirements.txt
selenium>=4.0.0
allure-pytest>=2.9.0
pytest>=6.0.0
pytest-html>=3.0.0